Further Refactoring
