+++ closetr-api/api/routes/clothes/clothes.controller.js
