+++ closetr-api/api/routes/clothes/clothes.model.js
