+++ closetr-api/api/routes/outfit_entries/outfit_entries.controller.js
