+++ closetr-api/common/result_handling.js
