+++ closetr-api/config/aliases.js
@@ -1,5 +1,6 @@
 const moduleAlias = require('module-alias');
 const path = require('path');
 
-moduleAlias.addAlias('@common', path.join(__dirname, '/../components/common'));
-moduleAlias.addAlias('@components', path.join(__dirname, '/../components'));
+moduleAlias.addAlias('@common', path.join(__dirname, '/../common'));
+moduleAlias.addAlias('@routes', path.join(__dirname, '/../api/routes'));
+moduleAlias.addAlias('@api', path.join(__dirname, '/../api'));
